backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio==5.11.3
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
pySmartDL==1.3.4
litellm==1.52.0